restructure react files, clean up code, comment out unused routes on backend, add env.example files
