+++ zoningtool-backend/.env.example
@@ -1,0 +1,9 @@
+PRIMARY_DB_URL=
+PRIMARY_DB_USERNAME=
+PRIMARY_DB_PASSWORD=
+
+SECONDARY_DB_URL=
+SECONDARY_DB_USERNAME=
+SECONDARY_DB_PASSWORD=
+
+APP_URL=
