+++ zoningtool-frontend/src/components/ZoningMap/styles/ZoningMap.css
@@ -1,0 +1,265 @@
+
+.legend {
+  padding: 12px 15px;
+  background: white;
+  border-radius: 5px;
+  box-shadow: 0 0 15px rgba(0, 0, 0, 0.2);
+  transition: all 0.3s ease;
+}
+
+.legend.minimized {
+  padding: 8px 15px;
+}
+
+.legend h4 {
+  margin: 0;
+  font-size: 14px;
+  font-weight: 600;
+  color: #5f6368;
+  display: inline-block;
+}
+
+.legend-item {
+  display: flex;
+  align-items: center;
+  margin-bottom: 5px;
+}
+
+.legend-item i {
+  width: 18px;
+  height: 18px;
+  margin-right: 8px;
+  opacity: 0.7;
+  display: inline-block;
+}
+
+.leaflet-control-custom:hover {
+  background-color: #f0f0f0 !important;
+}
+
+.legend-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+}
+
+.minimize-btn {
+  cursor: pointer;
+  font-size: 16px;
+  font-weight: 500;
+  margin-left: 10px;
+}
+
+.minimize-btn:hover {
+  color: #666;
+}
+
+.legend-content {
+  transition: all 0.3s ease;
+  margin-top: 10px;
+}
+
+/* Remove default margins and padding from body and html */
+body,
+html {
+  margin: 0;
+  padding: 0;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+  /* This prevents scroll bars */
+}
+
+.stats-box {
+  transition: all 0.3s ease;
+}
+
+.stats-box h3 {
+  color: #333;
+  font-size: 16px;
+  border-bottom: 1px solid #eee;
+  padding-bottom: 8px;
+}
+
+.stats-box p {
+  margin: 8px 0;
+}
+
+body {
+  font-family: 'Roboto', sans-serif;
+  color: #202124
+}
+
+.menu-item {
+  display: flex;
+  justify-content: space-between;
+  cursor: pointer;
+  align-items: center;
+  font-size: 19px;
+  margin-left: 10px;
+  margin-right: 15px;
+  font-weight: 500;
+}
+
+.section-divider {
+  border: none;
+  border-top: 1px solid #ccc;
+  margin: 25px 0;
+}
+
+.menu-position {
+  margin-top: 35px;
+}
+
+
+
+
+/* ParcelSelection.css */
+
+.parcel-selection-container {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.selection-mode-button {
+  background-color: #4CAF50;
+  color: white;
+  padding: 8px 16px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+.selection-mode-button.selection-active {
+  background-color: #f44336;
+}
+
+.selection-mode-button:hover {
+  background-color: #388E3C;
+}
+
+.selection-mode-button.selection-active:hover {
+  background-color: #d32f2f;
+}
+
+.selection-actions-row {
+  display: flex;
+  gap: 8px;
+  width: 100%; /* Ensure the row takes full width */
+}
+
+.selection-button {
+  background-color: #2196F3;
+  color: white;
+  padding: 8px 16px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+  flex: 1; /* This makes buttons take equal space */
+  text-align: center; /* Center the text */
+}
+
+.selection-button:hover {
+  background-color: #1976D2;
+}
+
+.selected-count {
+  margin: 8px 0;
+}
+
+.update-button {
+  background-color: #4CAF50;
+  color: white;
+  padding: 8px 16px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  width: 100%;
+  margin-top: 0px;
+  transition: background-color 0.3s;
+}
+
+.update-button:hover {
+  background-color: #388E3C;
+}
+
+.update-button:disabled {
+  opacity: 0.5;
+}
+
+.delete-button {
+  background-color: #f44336;
+  color: white;
+  padding: 8px 16px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  width: 100%;
+  margin-top: 10px;
+  transition: background-color 0.3s;
+}
+
+.delete-button:hover {
+  background-color: #d32f2f;
+}
+
+.delete-button:disabled {
+  opacity: 0.5;
+}
+
+.selected-parcels-list {
+  margin-top: 10px;
+  max-height: 200px;
+  overflow-y: auto;
+}
+
+.zoning-type-select {
+  margin: 10px 0;
+}
+
+.zoning-type-select label {
+  display: block;
+  margin-bottom: 5px;
+  font-weight: bold;
+}
+
+.zoning-type-select select {
+  width: 100%;
+  padding: 5px;
+}
+
+.side-panel {
+  position: absolute;
+  top: 0;
+  left: -340px;
+  width: 300px;
+  height: 100vh;
+  background-color: #fff;
+  box-shadow: 2px 0 5px rgba(0,0,0,0.3);
+  z-index: 999;
+  padding: 20px;
+  transition: left 0.3s ease-in-out;
+  box-sizing: border-box;
+}
+
+.side-panel.open {
+  left: 0;
+}
+
+.panel-close-btn {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  background-color: transparent;
+  border: none;
+  font-size: 20px;
+  cursor: pointer;
+}
+
+.panel-close-btn:hover {
+  color: #666;
+}
