+++ zoningtool-frontend/src/components/ZoningMap/utils/constants.js
@@ -1,0 +1,7 @@
+export const ZONING_COLORS = {
+  Residential: "#3bb2d0",
+  Commercial: "#f2a25c",
+  Planned: "#e55e5e",
+  Updated: "purple",
+  default: "#ccc",
+};
